backend/pyproject.toml
[tool.poetry]
authors = ["Admin <rvosistemas@outlook.com>"]
description = "Prueba softgic"
name = "app"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.12.1"
bcrypt = "4.0.1"
email-validator = "^2.1.0.post1"
emails = "^0.6"
fastapi = "^0.109.1"
gunicorn = "^22.0.0"
httpx = "^0.25.1"
jinja2 = "^3.1.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg = {extras = ["binary"], version = "^3.1.13"}
pydantic = "2.8.2"
pydantic-settings = "2.4.0"
pyjwt = ">=1.7.1,<2.0.0"
python = "^3.10"
python-multipart = "^0.0.7"
sentry-sdk = {extras = ["fastapi"], version = "^1.40.6"}
sqlmodel = "^0.0.21"
tenacity = "^8.2.3"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.3"
mypy = "^1.8.0"
pre-commit = "^3.6.2"
pytest = "^7.4.3"
ruff = "^0.2.2"
types-passlib = "^1.7.7.20240106"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool.mypy]
exclude = ["venv", ".venv", ".venv_*", "alembic"]
strict = true

[tool.ruff]
exclude = ["alembic"]
target-version = "py310"

[tool.ruff.lint]
ignore = [
  "E501", # line too long, handled by black
  "B008", # do not perform function calls in argument defaults
  "W191", # indentation contains tabs
  "B904", # Allow raising exceptions without from e, for HTTPException
]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
  "ARG001", # unused arguments in functions
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
